requirements.txt
torch
torchvision
torchaudio
numpy
pandas
scipy
scikit-learn
matplotlib
seaborn
huggingface-hub
transformers
fastapi
uvicorn
onnxruntime
tqdm
pillow
opencv-python
protobuf
pyyaml
pydantic
